Don't add implicit ctor/dtor for .con/destructor( ) method call
This reverts c5892a9419ad20d3cfccebb5512283defa94eac3 and closes #3524323.

Adding a constructor or destructor to an UDT after the TYPE ... END TYPE
was parsed is not safe because the compiler only does a single pass.

It might work in practice because it won't matter whether the implicitly
added ctor/dtor is getting called or not, but it can confuse many parts
of the compiler or create "impossible" situations, where first there was
no ctor, and then all of sudden there is one, even though it's still the
same UDT.

For example: It's not allowed to have a field named "constructor" in
an UDT that has methods. However, by using a ".constructor( )" method call
a constructor could still be added. Now, it's no longer clear whether
".constructor" refers to the field or the constructor.

As seen with bug #3524323, the compiler apparently cannot handle creating
methods for an UDT while parsing inside a method of the UDT. It is probably
based around the assumption that method bodies can only be defined outside
TYPE ... END TYPE, and no new methods will be declared after that, or
perhaps it's just that procedure nesting isn't handled well (except for
switching back and forth between the implicit main() proc and the others).
Either way, if ".constructor( )" doesn't auto-create constructors anymore,
then the condition for the compiler crash is gone.