parser: Disallow EXIT FUNCTION after RETURN with complex UDT result
This patch makes EXIT FUNCTION have properties similar to FUNCTION=.

For a function with a complex UDT result (i.e. an UDT with a constructor),
the result must be constructed at some point, before returning.

For FUNCTION=, this is done at the top of the function, in preparation for
the FUNCTION= statements that are coming.

RETURN on the other hand uses the copy-constructor to construct the result
right in place of each RETURN statement, like C++.

These two methods are not compatible to each other, and thus FUNCTION=
and RETURN are not allowed to be used together in functions returning
an UDT with a constructor.

This patch makes EXIT FUNCTION trigger the same behaviour that FUNCTION=
triggers, i.e. constructing the result at the top of the function.
EXIT FUNCTION will now be disallowed in combination with RETURN in such
functinos. This fixes the issues of EXIT FUNCTION leaving the result
unconstructed in "RETURN" functions, and also disallows RETURN after
an EXIT FUNCTION was seen.