ast: Use ANY PTR cast to initialize static pointers (fixes #3544952)
This works around issues with derived UDT pointers, which aren't allowed
to be cast to/from other pointer types. This can happen for example with
the data ptr field of static array descriptors if its for an array of
derived UDT pointers where the UDT was a forward reference.

For example:
	type Parent
		a as integer
	end type

	type Child as FwdrefChild

	dim shared array() as Child ptr

	type FwdrefChild extends Parent
	end type

	'' This REDIM ensures the array is used/referenced, forcing the static array
	'' desciptor to be emitted
	redim array(0 to 0)

	sleep

This caused the astNewCONV() in hFlushExprStatic() to fail when
initializing the array descriptor data field with UDT PTR PTR type with
an expression with FWDREF PTR PTR type (ASTNODEs aren't updated like
symbols are when FWDREFs become known). Since this happened during
emitting of globals, the type mismatch error was shown on unrelated
statements at EOF.