ast: Propagate CTORCALL type changes up to temp VAR access (fixes #3538470)
CTORCALL nodes apparently must be handled similar to FIELD nodes in this
regard. For example:

	type UDT
		i as integer
		declare constructor( )
	end type

	dim as integer x
	x = (UDT( )).i

The last assignment's rhs is a field access on a CTORCALL expression.
Since the field access is at offset 0 in the temp UDT var created by the
CTORCALL, it is optimized out, and the UDT var access expression (which
happens to be a CTORCALL here) is changed over to use the field's dtype.

This type change must be handled by the CTORCALL one way or another.
FIELD nodes handle it by propagating it up to the actual field access
(usually either a DEREF or VAR node), and it makes sense for the CTORCALL
to do the same thing: propagate the type change up to the expression
that it represents.

CTORCALLs contain two sub-trees:
	l = the ctor CALL on the temp UDT var,
	r = the VAR access to the temp UDT var.

The ctor CALL can be left alone, but the VAR access is what needs to be
updated; since this is the real expression represented by the CTORCALL.

Instead of handling this in astSetType(), the VAR access could also be
updated during astLoadCALLCTOR(), as long as the vreg ends up having the
correct type one way or another, but doing it in astSetType(), just like
FIELD nodes, seems to be much nicer and safer. At least then we know the
tree is correct and further optimizations won't make false assumptions.

Previously with the code snippet given above the vreg returned by the
astLoadCALLCTOR() had FB_DATATYPE_STRUCT type (because the VAR wasn't
updated to the field's FB_DATATYPE_INTEGER type), causing an out-of-bounds
array access to regTB() in ir-tac.bas(2104):hFlushStore(), because the
ASM emitter doesn't expect vregs to be structs.