C backend: Work-around local var shadowing
The problem is that FB code such as this:

	dim as integer i = 1
	scope
		print i
		dim as integer i = 2
		print i
	end scope

resulted in C code such as this:

	int I$;
	I$ = 1;
	{
		int I$;
		fb_PrintInt( I$ );
		I$ = 2;
		fb_PrintInt( I$ );
	}

where the inner "I$" variable shadows the outer one because it's declared
too early.

This happened because all local symbols were emitted at top of the scope
by astScopeAllocLocals() as for the ASM backend. Now, for the C backend,
local variables are emitted based on their AST DECL nodes, which results
in C code such as:

	int I$;
	I$ = 1;
	{
		fb_PrintInt( I$ );
		int I$;
		I$ = 2;
		fb_PrintInt( I$ );
	}

An alternative approach is to append the scope level to the variable's
name, in order to get unique variable names:

	int I0$;
	I0$ = 1;
	{
		int I1$;
		fb_PrintInt( I0$ );
		I1$ = 2;
		fb_PrintInt( I1$ );
	}

The DECL-based approach should work fine for the moment, but using the
scope levels would be safer. However, "i" becoming "I$" is bad enough,
adding the scope level would make it even worse (e.g. debugging with gdb).