C backend: Emit statics with dtors as globals with unique name
Static variables with destructors register a destructor wrapper function
with atexit() when first initialized, in order to clean themselves up
at program exit. The destructor wrapper calls the static var's destructor
on the static var.

For that to work, the wrapper proc of course needs to "see" the static var.
In FB or C code, static vars are not visible to other functions.
The C backend needs to work around this issue by emitting static vars
with dtor as global vars. For that symb-mangling will now give such vars
a unique name when using the C backend.

For the ASM backend this is less of a problem, because all the symbols are
globally visible anyways.

Accordingly, the C backend now uses irProcAllocStaticVars() again to emit
the statics with dtors. Everything else is still handled by astLoadDECL()
and astScopeAllocLocals().