Add local procptr protos to scope instead of namespace
This will prevent local procptr protos from being
re-used by other scopes. Adding them to the current
namespace was not safe, because they could reference
local symbols (e.g. UDTs), and besides that it caused
problems with the C backend.

Adding them to the current scope (when inside one) is
ok though, as no variable/type declared in a scope can
be used outside of it, so this should not break any
type checks involving procptrs.

25a36320 tried to fix this already, though that idea was
wrong and had to be reverted in 58dae4d3.