C backend: Don't worry about preventing typeAddrOf() overflows
dtypes use 4 bits for the pointer count, that allows 0 to 15, while the
soft-limit is 8. So, the backend can do extra typeAddrOf()s without
worrying about overflows.

Note that there only are 8+1 CONST bits, not 15+1, so CONST would in
theory be overflown when going over the 8 PTRs via extra typeAddrOf(),
but it looks like this won't happen in practice, at least not in the
C backend, because the CONST bits are never used there anyways.
Neither the C backend code nor IR vregs in general use "full" types;
the CONST bits should matter to the FB parser only.