LLVM backend: Add vreg accesses + offsets/loads/casts and proper stores
Similar to the ASM backend, VAR vregs must be loaded into a register first,
to get the variable's value from memory. In the LLVM backend the registers
are just REG vregs though, not x86 registers.

Also, operations that want to store their result into memory, are using
a temporary REG vreg as result now, and then STORE it into memory in a
separate operation (that's how it must be done in LLVM, unlike x86 ASM).