Declare implicit members before vtable creation, and implement afterwards
A dtor implemented by the user can override a virtual dtor from the base
class, the same must happen for an implicitly generated dtor, to ensure
it gets called.

For this to work the implicit dtor must be declared before building the
vtable, to prevent that code from thinking that there is no dtor for
that class.

However we can also not just declare and implement all implicit members
before building the vtable, because the bodies of implicit ctors require
the vtable to be known (for the this.vptr = @vtable+8 initialization).

Now, all implicit members are declared before vtable creation, and their
bodies are implemented afterwards. This allows implicit dtors to override
virtual base dtors.