makefile: Add 2-step C/ASM boostrapping commands
1. Precompile the compiler .bas source code files into .asm files,
resulting in a source tree containing these .asm files, which could be
packaged up as a special FB source package for bootstrapping:

	make prepare-asm-bootstrap

2. Build such an FB source tree, i.e. the compiler is built from the
precompiled .asm files, instead of its .bas sources, while the rtlib
and gfxlib2 are built normally. Due to the precompilation, it is not
necessary to have a working fbc binary for this step.

	make ENABLE_ASM_BOOTSTRAP=1

The same pattern applies to "make prepare-c-bootstrap" and
"make ENABLE_C_BOOTSTRAP=1", except that these use -gen gcc and
precompiled .c files.

This functionality may be interesting for making .deb/.rpm source packages,
which typically cannot rely on precompiled binaries, thus using
precompiled .asm or .c files may be the only option to get fbc packaged.