Disallow types with nested anonymous unions in LET() and THREADCALL
Since unions were already disallowed there, structs with nested anonymous
unions should be disallowed too, since they're effectively the same.

Especially LET() can not know which field of the union to assign to which
variable. Only the first? or all? Neither makes much sense or is useful.

Types with fields AS UnionUDT (non-anonymous nested unions) should still
be allowed though, at least with LET, because they are copied to UnionUDT
variables, i.e. the whole union is copied, not just a part of it.