rtlib: Print error messages to stderr without waiting for console reset
I'd like to fix up Assert() and AssertWarn() to print directly to
stderr, instead of using the normal FB PRINT mechanism, because that
goes to the FB SCREEN - but printing an assert() message to a graphics
screen when it's going to be closed immediately afterwards is useless.

Using fputs() to stderr seems to work fine in testing, however I found
that the runtime error messages originally did the same but then were
changed to store the error message in a temporary buffer in fb_Die()
and let fb_hRtExit() show the error after resetting the rtlib's
console I/O business. (since 8c97f1ad)

That raised the question whether it'd be "safe" to print to stderr
directly. Unfortunately I do not know what reasons there were/are for
delaying the printing to stderr until after the console stuff was
cleaned up. Testing suggests it is unnecessary nowadays though.

Since it seems like it's not needed anymore, and the rtlib is not
touching stderr at all on its own, and the gfxlib does not reset
the console on exit, and we already use fputs directly on stdout
to print Unicode strings/chars, the same should work with stderr.

Then, Assert() and AssertWarn() can do the same thing as fb_Die().