rtlib: Print assert() messages to stderr, like runtime errors (fixes #637)
This changes Assert() and AssertWarn() to print directly to stderr
instead of using the normal FB PRINT mechanism. This means the output
will always appear in the console (or whereever stderr was redirected
to), but it also means that the assertion messages will be visible even
when using the FB graphics mode.

Printing the assertion messages to the FB screen directly is useless
because the assert() will also exit the program causing the gfxlib2
to close the screen window immediately, preventing the user from seeing
the message at all.

Runtime error messages were already printed to stderr, so it seems good
to do the same for assertions.