Disallow creating objects of abstract classes
After this patch, UDTs with unimplemented methods can no longer be
used directly with DIM/NEW etc., i.e. no variables/objects of such a
type can be created directly anymore. As in C++, only subclasses that
implement all the abstract methods can be used to create objects.

There is no technical reason for this (at least currently even an UDT
with only ABSTRACT methods would still extend OBJECT and thus contain
the vptr field, i.e. would not really be empty), but the idea is that
allowing it does not make sense in terms of design:

If a class has ABSTRACTs, we assume that it intends to use them,
in fact that it *does* use them, and now the ABSTRACTs must be
implemented, because they are used (or else the program would
call unimplemented abstract methods, which we currently handle as
NULL ptr crash).

What would be the point of having unused ABSTRACTs in a class?
From the OOP/static typing point of view an abstract class is an
interface which is a contract. A subclass can implement it or not,
but a partial implementation does not make sense. How should other
code know which methods are implemented or not? Where a partial
implementation seems needed, the interface should really be split
up into two, most likely with one extending the other, then we can
again rely on the static typing to be safe.

Finally, with objects of abstract classes disallowed, we can
effectively assume that ABSTRACT methods are always implemented,
because only objects of classes with all ABSTRACTs implemented can
be used (the only exception is while executing an ABSTRACT class
constructor, before the subclass updates the vptr...). Then the problem
of how to handle unimplemented ABSTRACTs (almost) goes away...