ASM backend: Remove unsigned MUL code and use IMUL instead
Even though the buggy MUL code generation for unsigned 16/32bit integer
multiplication was fixed in 8aba5fcc, it's not worth it to keep it around
when we can use IMUL as for signed multiplication and get better code,
just like C compilers. The only advantage of MUL is that it preserves
the upper 32bits of the result (in EDX), but we didn't use that anyways.

See also:
http://www.freebasic.net/forum/viewtopic.php?p=178706#p178706
http://stackoverflow.com/a/4040834