Don't bother omitting struct tail padding for SWAP/MEMMOVE/MEMCLEAR
The tail padding rounds up the struct size, which only makes the operation
faster since it can work on whole dwords and doesn't have to work through
any 1, 2 or 3 byte remainder.

Besides that, the tail padding bytes are part of the struct (they are
included in the sizeof()) and thus should never be ignored,
e.g. they could be important when doing file I/O or when hashing the
struct byte-by-byte...