Add symbMakeParamOptional(), to add initializers after creating the params
This allows the parameter declaration parser to add the param symbol
before parsing an initializer expression; that way it can check the
expression against the parameter symbol directly.

This simplifies the parsing code a bit, and will allow cInitializer()
to be called on the param symbol directly (instead of having to use a
temp var) in a follow-up patch.