param parser: Use cInitializer() also for non-UDT param initializers
Similar to 53c39679, we must use cInitializer() in order to get a TYPEINI
tree with a temp scope that captures any temp symbols allocated by the
initializer expression. It does not matter whether it's a UDT or non-UDT
param, the same problems with temp symbols can appear. Also, field
initializers do not make this distinction either.