Don't make temp vars STATIC in procs with STATICLOCALS
It's not safe; STATIC temp vars will break things like FOR loops or
SELECT CASE under multi-threading or even WITH blocks when the proc is
recursive.

The problem is that temp vars are hidden, the programmer never knows that
certain constructs suddenly became unsafe when STATICLOCALS was used.
So it's better to be consistent and never make temp vars STATIC,
STATICLOCALS should only apply to vars DIM'ed by the programmer.