Use symbGetProcRealType() more often, instead of remapping manually
which allows simplification, especially in astLoadCALL().

Also, this patch adds a corresponding symbGetProcRealSubtype() in order
to ensure the correct subtype is used together with the real dtype.
(for example, the real subtype will be NULL if a proc returns an UDT in
registers with dtype = INTEGER, even though the proc's subtype is an UDT)