Fix bad cloning of the new[] ctorlist loop code
new[] uses a loop to call the default constructor of each element (if the
type has a defctor). The loop of course uses a label, i.e. the built up
tree includes a LABEL (that declares the label at the top of the loop)
and a corresponding BOP/BRANCH node (that does the jump back up).

The problem is that new[] is an expression, and the initialization loop
is included in the tree by using LINK nodes, instead of being astAdd()'ed
immediately like all (really?) other cases where the compiler builds up
loops like that.

This patch essentially fixes astCloneTree() to handle this case properly.
Even though it's rare, there can be cases when the compiler will clone
an expression tree containing a new[] and its loop.

Now, the new[] loop will be wrapped in the new LOOP node, allowing
astCloneTree() to identify this case. It will duplicate the label symbol,
and then update the copy of the original tree accordingly, especially
the LABEL and BOP/BRANCH nodes, ensuring the cloned new[] loop will no
longer use the same label like the original new[] loop.
(This is similar to how iif()'s falselabel is handled)

Thus astCloneTree() on such new[] expressions won't cause duplicate labels
and bad jumps to be emitted anymore.