Use non-flushing LABEL for the new[] ctorlist loop
The new[] loop code will be part of an expression tree, so the LABEL node
cannot be emitted as a flushing label, that would cause the ASM emitter
to flush() all live expressions, try to reuse vregs, etc.

Instead, the non-flushing version must be used, so the whole expression
tree can be emitted safely (without vregs being overwritten).