Fix missing +BOP -> WstrConcat updating in wstring sidefx removal
For example, SELECT CASE wstring-expression uses astRemSideFx() if
the wstring-expression contains some CALL (i.e. has side-effects),
because it wants to generate:

	selectwcharptr = WstrAlloc( len( wstring-expression ) )
	WstrAssign( selectwcharptr, wstring-expression )

but if there are side-effects it must become:

	wcharptr = @wstring-expression
	selectwcharptr = WstrAlloc( len( *wcharptr ) )
	WstrAssign( selectwcharptr, *wcharptr )

The sidefx-removal temp var is a wchar ptr because that's how
astRemSideFx() handles all string types and also UDTs. Using a reference
is more efficient than creating a full copy, and it may actually even be
necessary because later taking the address of the reference will give the
correct result, while taking the address of a full copy would return the
address of the full copy, not the original expression.

In case there are no side-effects, the ARG nodes in the WstrAlloc() and
WstrAssign() CALLs invoke astUpdStrConcat() and all is well

In case there are side-effects, that's no longer the case, because the
argument to the WstrAlloc() and WstrAssign() is just the *wcharptr,
not the original wstring-expression.

On top of that, astOptimize() will not handle the wcharptr assignment:
	wcharptr = @wstring-expression
because it only handles wchar/wstring assignments as in:
        wstring = wstring-expression
so astUpdStrConcat() won't be called from there either.

This patch adds the astUpdStrConcat() call to astRemSideFx(), which is the
code that's taking a string expression and turning it into an assignment,
which however is not a string assignment, but just a pointer assignment,
so it won't be handled by astOptimize().

The lack of +BOP -> WstrConcat updating resulted in bad code being
generated (causing the program to crash at runtime). In theory the
bug could show up with any constructs using astRemSideFx(), but in
practice almost all string expressions will either end up in string
assignments (handled by astOptimize()) or as string arguments (handled
by ARG nodes). SELECT CASE might just be the only case where the bug
could show up...