ast: Fix astNewCONSTz() to handle structs correctly
And fix the callers to pass the subtype too.

It will now create a nice TYPEINI tree, i.e. a legal AST with
FB_DATATYPE_STRUCT, instead of returning a CONST( 0 ) with STRUCT PTR
type. This prevents the assert() on the astNewASSIGN() result in
ast-node-typeini.bas:hFlushTree() from triggering during error recovery.