ASM backend: Fix unnecessary MOV when preserving LONGINT regs across CALLs
LONGINT vregs use two registers, the main vreg for the low dword and an
auxiliary vreg for the high dword.

In hPreserveRegs() vregs associated with caller-saved registers are copied
to callee-saved registers before CALLs, via an emitMOV(). This is intended
to copy registers, not vregs. The difference is that in case of LONGINTs
which use two regs, only one may need to be copied, if the other already
is in a callee-saved register.

But hPreserveRegs() didn't handle this case, it just invoked emitMOV()
which then copied two registers when it was only intended to copy one.
Due to
	tr = *vr
the tr vreg had the same IRVREG->vaux field as the vr vreg,
thus emitMOV() would always generate a redundant MOV for it.

Now hPreserveRegs() will update the temporary vregs given to the emitMOV()
to prevent that from happening.