Fix SELECT CONST with only CASE ELSE, triggered assert() since af46737d
The jump table can be empty (labelcount = 0), for example:
	SELECT CASE AS CONST ...
	CASE ELSE
	END SELECT

or even during error recovery from:
	SELECT CASE AS CONST ...
	END SELECT

Now the compiler will simply emit a jump instead of a jump table and the
corresponding checks on the SELECT expression. The SELECT temp var will
end up being unused (which is why the GCC backend takes special care to
work-around gcc "var set but not used" warnings). Either way, the generated
code is nicer now for this specific case.