Fix bad temp var dtor calls for the unreached expression in an iif()
Temp vars used by the true-expression or the false-expression in an iif()
will only be constructed when their expression is executed. Since the whole
point of iif() is execute only one of the two, we cannot just destruct
all temp vars at the end of the whole expression that included the iif().

Instead, essentially we must destruct only the temp vars that were
constructed, i.e. whose code path was chosen in the iif(). The easiest
way to do this is to emit the dtor calls for these temp vars right into
the iif()'s true or false code paths respectively.

Another way to do this would be emit a boolean "is_constructed" flag for
each temp var (could be done in form of an integer temp var, or even bit
flags to avoid having to allocate tons of flag temp vars if there are lots
of temp vars that need destructing).

On the other hand, saying that each temp var will always be constructed
(so that we could always safely & easily destruct it) would be wrong,
because expressions that are not reached in an iif() should not be
executed, not even temp var ctor calls.

Thus, it seems like the first solution (emitting the dtor calls into the
true/false code paths of the iif()) is the best way to go. It would be
risky to do this if the temp vars from the true/false expressions could
be used/referenced outside the iif() somehow, but currently I can't see
that happening ever, especially since iif() uses a temp var itself to
store the result. As long as we emit the dtor calls behind the assignments
to the iif temp var, it should be safe.

Of course all this also affects the ANDALSO/ORELSE short-circuiting
operators since their second operator can also be an unreachable
expression. Thankfully they are implemented using iif() internally,
so fixing them too is easy.