Fix type() and type<>() type checks for built-in types
Even though someinteger = type( 123 ) worked, someprocptr = type( ... )
did not, same for enums, because the old checking code assumed that
if the parser.ctxsym is set, it's the subtype of a struct. But of course
other types such as function pointers and enums use a subtype too, so
that assumption wasn't right.

(see also: 5fa418b5f0c2ddad955a386671da1aca00fb92c6)