Don't copy construct temp var in BYREF UDT parameter initializers
Parameters can currently be implicitly constructed from the argument,
if a matching constructor exists. I.e. a temp var will be created, much
like when passing an integer literal number to a BYREF AS INTEGER param.

But - a temp copy must not be created when the argument expression itself
is valid to be passed BYREF, because otherwise the BYREF semantics would
be destroyed (due to the temp var it would behave like BYVAL).

astNewARG() usually handles all this directly, it won't try hImplicitCtor()
when passing BYREF as-is is ok.

This patch fixes cInitializer() which is used for parameter initializers
to behave the same. For example, assuming UDT has a copy constructor:

	dim shared x0 as UDT
	declare sub passbyref( byref x as UDT = x0 )

cInitializer() produced a TYPEINI( TYPEINI_CTORCALL ), i.e. a copy
constructor call. And the TYPEINI will be turned into a temp var later,
destroying BYREF semantics, because what should astNewARG() due with such
a TYPEINI? It would have to check whether it's a copy constructor call,
and then extract the original expression... it's much easier to fix up
cInitializer() and let it produce a more simpler TYPEINI( TYPEINI_ASSIGN )
for that case, which astNewARG() can unwrap using astTypeIniTryRemove().