Don't initialize EXTERN array descriptors (fixes #630)
EXTERN array descriptors should not be given an initializer, because they
are extern, and they will be initialized in another module already.

This only affected the C backend though (and LLVM perhaps), which got
confused due an EXTERN having an initializer, causing it to assume it
wasn't an EXTERN. The ASM backend checks for the EXTERN attribute first,
so was not affected.