parser: Remove unnecessary return values of various statement parsers
Anything that handles a statement (by parsing it or by reporting an error
and doing error recovery) doesn't need to return TRUE/FALSE to
cDeclaration() or cCompoundStmt(). If the statement was parsed, we're at
EOL, and cStatement() cannot continue trying anything else anyways.
The FALSE result in these cases is only needed if a parsing function did
not handle a certain construct, and wants to let another parsing function
handle it.